requirements.txt
# Core
flask==3.1.3
python-dotenv==1.2.2

# Excel processing
polars[excel]==1.39.3
openpyxl==3.1.5

# Production server
waitress==3.0.2

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.40
